Split custom sidebar pane view files
